reactive_graph_runtime_api::runtime

Trait Runtime

pub trait Runtime:
    TypeSystem
    + CommandSystem
    + ConfigSystem
    + GraphQLSystem
    + DynamicGraphSystem
    + RuntimeGraphQLSystem
    + PluginGraphQLSystem
    + RemotesSystem
    + PluginSystem
    + BehaviourSystem
    + InstanceSystem
    + ReactiveSystem
    + RuntimeSystem
    + WebSystem
    + Send
    + Sync
    + Lifecycle {
    // Required methods
    fn config<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn run<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stop(&self);
    fn is_running(&self) -> bool;
    fn wait_for_started<'life0, 'async_trait>(
        &'life0 self,
        timeout_duration: Duration,
    ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_for_stopped<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_for_stopped_with_timeout<'life0, 'async_trait>(
        &'life0 self,
        timeout_duration: Duration,
    ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn address(&self) -> InstanceAddress;
}

Required Methods

fn config<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn run<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn stop(&self)

fn is_running(&self) -> bool

fn wait_for_started<'life0, 'async_trait>( &'life0 self, timeout_duration: Duration, ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server to be started. Times out if the GraphQL server is not running after the given duration.

fn wait_for_stopped<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server has been stopped.

fn wait_for_stopped_with_timeout<'life0, 'async_trait>( &'life0 self, timeout_duration: Duration, ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server has been stopped. Times out if the GraphQL server is still running after the given duration.

fn address(&self) -> InstanceAddress

Returns the address of the runtime.